Tests/LibWeb/Text/expected/HTML/picture-source-dimension-attributes.txt

with source dimensions: 200x100
without source dimensions: 50x50
